config: the cuda-graph capture decisions move out of the record
Ten handlers, one family: parsing `--cuda-graph-*` into a `CudaGraphConfig`,
the compatibility rules that disable a capture backend, the DeepEP bucket
alignment, and the two per-architecture prefill defaults. They live in
`arg_groups/cuda_graph_hook.py` now and call each other directly; the record
keeps a call at each slot.

Moving them showed that two guards were reading the record's file and nothing
else, so a handler that leaves stops being watched:

- `test_model_config_reads_resolved_input` pins the opaque plugin callbacks
  against the first `get_model_config()` build. Its walk stopped at the slot,
  found a later build, and failed. It now follows the slot into `arg_groups/`.
- `test_resolution_reads_no_bag` seeded its reachability walk from
  `server_args.py` and `overrides.py`. It now seeds from the whole package.

62 launch shapes report the same outcome as before, and the 24-shape
resolution dump is field-for-field identical.
