+++ test/registered/unit/server_args/test_resolution_reads_no_bag.py
@@ -72,6 +72,15 @@ _ATTRIBUTE_SPELLED = _BAG_ACCESSORS - {"get_device"}
 _OWN = ("server_args.py", "runtime_context.py")
 
 
+def _pipeline_sources():
+    """The record plus every module under `arg_groups/`.
+
+    A handler that moved out of the record takes its imports with it, so
+    seeding the walk from two files would stop covering it.
+    """
+    return [_SRT / "server_args.py", *sorted((_SRT / "arg_groups").rglob("*.py"))]
+
+
 def _module_of(name):
     """`sglang.srt.a.b` -> the file, if it is one of ours."""
     if not name or not name.startswith("sglang.srt."):
@@ -241,9 +250,7 @@ class TestResolutionReadsNoBag(CustomTestCase):
 
     def test_the_walk_finds_something_to_walk(self):
         """A collapsed import map would make the pin vacuous."""
-        imported = _imported_symbols(
-            [_SRT / "server_args.py", _SRT / "arg_groups" / "overrides.py"]
-        )
+        imported = _imported_symbols(_pipeline_sources())
         self.assertGreater(
             len(imported),
             20,
@@ -282,9 +289,7 @@ class TestResolutionReadsNoBag(CustomTestCase):
         )
 
     def test_nothing_the_pipeline_calls_reads_a_bag(self):
-        imported = _imported_symbols(
-            [_SRT / "server_args.py", _SRT / "arg_groups" / "overrides.py"]
-        )
+        imported = _imported_symbols(_pipeline_sources())
         reachable = {
             (path, symbol) for path, symbols in imported.items() for symbol in symbols
         } | _registered_entries()
